+++ packages/cloud-cognitive/src/components/Datagrid/utils/DatagridActions.js
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React, { useLayoutEffect, useState } from 'react';
+import React, { useLayoutEffect, useState, useContext } from 'react';
 import {
   TableToolbarContent,
   TableToolbarSearch,
@@ -18,13 +18,21 @@ import {
   ModalFooter,
   Dropdown,
 } from '@carbon/react';
-import { Download, Add, Restart, ChevronDown } from '@carbon/react/icons';
+import {
+  Download,
+  Add,
+  Restart,
+  ChevronDown,
+  Filter,
+} from '@carbon/react/icons';
 import { action } from '@storybook/addon-actions';
 import { pkg } from '../../../settings';
 import { ButtonMenu, ButtonMenuItem } from '../../ButtonMenu';
+import { FilterContext } from '../Datagrid/addons/Filtering';
 
 const blockClass = `${pkg.prefix}--datagrid`;
 export const DatagridActions = (datagridState) => {
+  const { setPanelOpen } = useContext(FilterContext);
   const {
     selectedFlatRows,
     setGlobalFilter,
@@ -35,6 +43,7 @@ export const DatagridActions = (datagridState) => {
     filterProps,
     getFilterFlyoutProps,
     FilterFlyout,
+    data,
   } = datagridState;
   const downloadCsv = () => {
     alert('Downloading...');
@@ -56,6 +65,21 @@ export const DatagridActions = (datagridState) => {
       <FilterFlyout {...getFilterFlyoutProps()} />
     );
 
+  const renderFilterPanelButton = () =>
+    filterProps?.variation === 'panel' && (
+      <Button
+        kind="ghost"
+        hasIconOnly
+        tooltipPosition="bottom"
+        renderIcon={(props) => <Filter size={16} {...props} />}
+        iconDescription={filterProps.panelIconDescription}
+        className={`${blockClass}-filter-panel-open-button`}
+        onClick={() => setPanelOpen((open) => !open)}
+        disabled={data.length === 0}
+        tooltipAlignment="start"
+      />
+    );
+
   const [modalOpen, setModalOpen] = useState(false);
   const [size, setSize] = useState(window.innerWidth);
   useLayoutEffect(() => {
@@ -74,6 +98,7 @@ export const DatagridActions = (datagridState) => {
       <TableToolbarContent size="sm">
         {!mobileToolbar ? (
           <>
+            {renderFilterPanelButton()}
             <div style={style}>
               <Button
                 kind="ghost"
@@ -85,18 +110,17 @@ export const DatagridActions = (datagridState) => {
               />
             </div>
             {renderFilterFlyout()}
+            {CustomizeColumnsButton && (
+              <div style={style}>
+                <CustomizeColumnsButton />
+              </div>
+            )}
             <RowSizeDropdown {...rowSizeDropdownProps} />
             <div style={style} className={`${blockClass}__toolbar-divider`}>
               <Button kind="ghost" renderIcon={Add} iconDescription={'Action'}>
                 Ghost button
               </Button>
             </div>
-
-            {CustomizeColumnsButton && (
-              <div style={style}>
-                <CustomizeColumnsButton />
-              </div>
-            )}
           </>
         ) : (
           <OverflowMenu ariaLabel="Tools" size="md" flipped>
@@ -115,6 +139,7 @@ export const DatagridActions = (datagridState) => {
       </TableToolbarContent>
     ) : !mobileToolbar ? (
       <TableToolbarContent>
+        {renderFilterPanelButton()}
         <TableToolbarSearch
           size="xl"
           id="columnSearch"
@@ -123,7 +148,6 @@ export const DatagridActions = (datagridState) => {
           onChange={(e) => setGlobalFilter(e.target.value)}
         />
         {renderFilterFlyout()}
-        <RowSizeDropdown {...rowSizeDropdownProps} />
         <div style={style}>
           <Button
             kind="ghost"
@@ -149,6 +173,7 @@ export const DatagridActions = (datagridState) => {
             <CustomizeColumnsButton />
           </div>
         )}
+        <RowSizeDropdown {...rowSizeDropdownProps} />
         <ButtonMenu
           label="Primary button"
           renderIcon={ChevronDown}
@@ -170,6 +195,7 @@ export const DatagridActions = (datagridState) => {
       </TableToolbarContent>
     ) : (
       <TableToolbarContent>
+        {renderFilterPanelButton()}
         <TableToolbarSearch
           size="xl"
           id="columnSearch"
@@ -177,6 +203,7 @@ export const DatagridActions = (datagridState) => {
           placeHolderText={searchForAColumn}
           onChange={(e) => setGlobalFilter(e.target.value)}
         />
+        {renderFilterFlyout()}
         <OverflowMenu
           ariaLabel="Tools"
           size="lg"